feat(ui): split settings into network and serial ports pages
Add a dedicated Network view for Ethernet and Wi‑Fi configuration and a SerialPorts view for RS‑485 settings. Update routing to expose the new /network route and repurpose the existing /settings route for serial ports. Adjust sidebar icons, i18n messages, and RsSettings component to support the I/O bus sub‑section. Remove the old combined Settings view.

BREAKING CHANGE: the previous Settings page is removed; its functionality is now split between the new Network and SerialPorts pages.
