+++ main/frontend/src/i18n/messages.ts
@@ -3,7 +3,8 @@ export const messages = {
     connection_error: 'Connection error',
     save: 'Save',
     dashboard: 'Dashboard',
-    settings: 'Settings',
+    network: 'Network',
+    serial_ports: 'Serial ports',
     system: 'System',
     login: 'Username',
     password: 'Password',
@@ -24,7 +25,8 @@ export const messages = {
     connection_error: 'Ошибка подключения',
     save: 'Сохранить',
     dashboard: 'Обзор',
-    settings: 'Настройки',
+    network: 'Сеть',
+    serial_ports: 'Послед. порты',
     system: 'Система',
     login: 'Имя пользователя',
     password: 'Пароль',
@@ -45,7 +47,8 @@ export const messages = {
     connection_error: 'Қосылым қатесі',
     save: 'Сақтау',
     dashboard: 'Шолу',
-    settings: 'Баптаулар',
+    network: 'Желі',
+    serial_ports: 'Сериялық порттар',
     system: 'Жүйе',
     login: 'Пайдаланушы аты',
     password: 'Құпиясөз',
@@ -66,7 +69,8 @@ export const messages = {
     connection_error: 'Errore di connessione',
     save: 'Salva',
     dashboard: 'Panoramica',
-    settings: 'Impostazioni',
+    network: 'Rete',
+    serial_ports: 'Porte seriali',
     system: 'Sistema',
     login: 'Nome utente',
     password: 'Password',
@@ -87,7 +91,8 @@ export const messages = {
     connection_error: 'Verbindungsfehler',
     save: 'Speichern',
     dashboard: 'Übersicht',
-    settings: 'Einstellungen',
+    network: 'Netzwerk',
+    serial_ports: 'Serielle Schnittst.',
     system: 'System',
     login: 'Benutzername',
     password: 'Passwort',
